moved Yes3 methods into Yes3Trait
