+++ html/yes3.html
@@ -67,7 +67,7 @@
        </div>
     </div>
  
-    <div id="yes3-hello-message" class="yes3-panel-row yes3-panel-message yes3-ellipsis">
+    <div id="yes3-hello-message" class="yes3-panel-row yes3-panel-message">
     </div>
  </div> <!-- hello -->
